requirements.txt
opencv-python
numpy
evdev
pyusb
